third_party/blink/web_tests/platform/win/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-generated-content-expected.txt

This is a testharness.js-based test.
FAIL fieldset generated content assert_equals: clientWidth expected 57 but got 46
Harness: the test ran to completion.
